+++ summoner-cli/CHANGELOG.md
@@ -22,6 +22,8 @@ The changelog is available [on GitHub][2].
   Support GHC `8.6.5` in generated projects. Use GHC `8.6.5` in Summoner.
 * Remove support of GHC `8.6.4`, `8.6.3`, `8.4.3`. Leave only the latest
   versions of each major release.
+* [#338](https://github.com/kowainik/summoner/issues/338):
+  Build Summoner with GHC-8.8.1.
 
 ## 1.3.0.1 — Apr 10, 2019